added an option not to delete downloads

+++ scripts/bash/UKB_connectivity_mapping_pipeline.sh
@@ -36,6 +36,7 @@ NC='\033[0m'
 main_dir=$1
 ukb_subjects_dir=$2
 line_number=$3
+delete_download=$4
 
 working_dir=`pwd`
 
@@ -341,7 +342,7 @@ echo -e "${GREEN}[INFO]`date`:${NC} All fMRI time-series generated."
 
 # for now, we'll just delete all downloaded files as the downloads are rather fast
 # from the high performance computing servers
-if [ -d "${ukb_subjects_dir}/${ukb_subject_id}_${ukb_instance}" ]; then
+if [ -d "${ukb_subjects_dir}/${ukb_subject_id}_${ukb_instance}" ] && [ ${delete_download} != "no-delete" ]; then
 	rm -r "${ukb_subjects_dir}/${ukb_subject_id}_${ukb_instance}/"
 fi
